move out nep21-specific functionality into its own file
